# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TakaBachao ScamShield ML Service",
    "description": "Inference API providing TF-IDF char n-gram classification, probability calibration, scam type prediction, and n-gram attribution reason codes.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/predict": {
      "post": {
        "summary": "Predict V1",
        "operationId": "predict_v1_v1_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict": {
      "post": {
        "summary": "Predict Alias",
        "description": "Alias for /v1/predict",
        "operationId": "predict_alias_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "loaded": {
            "type": "boolean",
            "title": "Loaded"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "sklearn_version": {
            "type": "string",
            "title": "Sklearn Version"
          },
          "feature_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Count"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "loaded",
          "model_version",
          "threshold",
          "sklearn_version"
        ],
        "title": "HealthResponse"
      },
      "PredictRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "description": "Message text in Bengali, Banglish, or English"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "label_at_threshold": {
            "type": "integer",
            "title": "Label At Threshold"
          },
          "scam_type": {
            "type": "string",
            "title": "Scam Type"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          }
        },
        "type": "object",
        "required": [
          "probability",
          "label_at_threshold",
          "scam_type",
          "reason_codes",
          "model_version",
          "threshold"
        ],
        "title": "PredictResponse"
      },
      "ReasonCode": {
        "properties": {
          "ngram": {
            "type": "string",
            "title": "Ngram"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          }
        },
        "type": "object",
        "required": [
          "ngram",
          "contribution"
        ],
        "title": "ReasonCode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}